refacto categories and tags at profile page

+++ client/src/components/Categories/Categories.jsx
@@ -7,17 +7,15 @@ import { Modal } from '../Modal'
 import { getAll, createOne, updateOne, deleteOne } from '../../apis/categories'
 
 const Categories = () => {
+  const [isLoading, setIsLoading] = useState(true)
+
   const [categories, setCategories] = useState([])
   const [categoryName, setCategoryName] = useState('')
   const [categoryId, setCategoryId] = useState('')
 
   const [submitType, setSubmitType] = useState('')
-
-  const [isLoading, setIsLoading] = useState(true)
-
-  const [createCategory, setCreateCategory] = useState(false)
-  const [updateCategory, setUpdateCategory] = useState(false)
-  const [deleteCategory, setDeleteCategory] = useState(false)
+  const [modalType, setModalType] = useState('')
+  const [messages, setMessages] = useState([])
 
   const getCategories = async () => {
     const res = await getAll()
@@ -26,42 +24,46 @@ const Categories = () => {
     setIsLoading(false)
   }
 
-  const triggerModal = (type, params = '') => {
-    setSubmitType(type)
-    setCategoryId(params)
+  const triggerModal = (type, id = '', name = '') => {
+    setMessages([])
 
-    if (type === 'create') {
-      setCreateCategory(true)
-    } else if (type === 'update') {
-      setUpdateCategory(true)
-    } else if (type === 'delete') {
-      setDeleteCategory(true)
-    }
+    setModalType(type)
+    setSubmitType(type)
+    setCategoryId(id)
+    setCategoryName(name)
   }
 
   const handleSubmit = async (e) => {
     e.preventDefault()
+    const name = categoryName
 
-    try {
-      let res = {}
+    let message = []
 
-      if (submitType === 'create') {
-        res = await createOne({ name: categoryName })
-        setCreateCategory(false)
-      } else if (submitType === 'update') {
-        res = await updateOne(categoryId, { name: categoryName })
-        setUpdateCategory(false)
+    if (name.length === 0 && (submitType === 'create' || submitType === 'update')) message = [...message, 'Name cannot be empty']
+
+    if (message.length > 0) {
+      setMessages(message)
+    } else {
+      try {
+        let res = {}
+
+        if (submitType === 'create') {
+          res = await createOne({ name })
+        } else if (submitType === 'update') {
+          res = await updateOne(categoryId, { name })
+        } else if (submitType === 'delete') {
+          res = await deleteOne(categoryId)
+        }
+
+        setMessages([res.data.message])
+        setModalType('')
+        setSubmitType('')
+        setCategoryName('')
         setCategoryId('')
-      } else if (submitType === 'delete') {
-        res = await deleteOne(categoryId)
-        setDeleteCategory(false)
-        setCategoryId('')
+        getCategories()
+      } catch (err) {
+        console.error(err)
       }
-
-      alert(res.data.message)
-      getCategories()
-    } catch (err) {
-      console.error(err)
     }
   }
 
@@ -72,13 +74,13 @@ const Categories = () => {
   return (
     <>
       <div className='mb-3 d-flex justify-content-between'>
-        <h2 className='fw-bold fs-3 d-inline mb-0'>LIST OF CATEGORIES</h2>
+        <h2 className='fw-bold fs-3 d-inline mb-0 text-uppercase'>list of categories</h2>
 
         <Button
-          className='text-light py-0 px-3'
+          className='text-light py-0 px-3 text-uppercase'
           onClick={() => triggerModal('create')}
         >
-          ADD NEW CATEGORIES
+          add new categories
         </Button>
       </div>
 
@@ -91,22 +93,22 @@ const Categories = () => {
           {categories.map((tag) => (
             <div key={tag._id}>
               <div className='d-flex justify-content-between p-3'>
-                <div className=''>
-                  <h5 className='mb-2 text-muted'>Name:</h5>
+                <div>
+                  <h5 className='mb-2 text-muted text-capitalize'>name:</h5>
 
                   <h3 className='m-0 fs-4 fw-bold'>{tag.name}</h3>
                 </div>
 
                 <div className='d-flex'>
                   <Button
-                    onClick={() => triggerModal('update', tag._id)}
+                    onClick={() => triggerModal('update', tag._id, tag.name)}
                     className='bg-transparent align-self-center me-2'
                   >
                     <FiEdit className='fs-5 text-dark' />
                   </Button>
 
                   <Button
-                    onClick={() => triggerModal('delete', tag._id)}
+                    onClick={() => triggerModal('delete', tag._id, tag.name)}
                     className='bg-transparent align-self-center'
                   >
                     <MdDeleteForever className='fs-5 text-dark' />
@@ -120,28 +122,46 @@ const Categories = () => {
         </>
       )}
 
-      <Modal
-        trigger={createCategory}
-        setTrigger={setCreateCategory}
-        type={'category'}
-        setName={setCategoryName}
-        submit={handleSubmit}
-      />
-      <Modal
-        trigger={updateCategory}
-        setTrigger={setUpdateCategory}
-        type={'category'}
-        isUpdate
-        setName={setCategoryName}
-        submit={handleSubmit}
-      />
-      <Modal
-        trigger={deleteCategory}
-        setTrigger={setDeleteCategory}
-        type={'category'}
-        isDelete
-        submit={handleSubmit}
-      />
+      {modalType === 'create' ? (
+        <Modal
+          trigger={modalType === 'create'}
+          setTrigger={setModalType}
+          type={'category'}
+          setName={setCategoryName}
+          submit={handleSubmit}
+          messages={messages}
+        />
+      ) : modalType === 'update' ? (
+        <Modal
+          trigger={modalType === 'update'}
+          setTrigger={setModalType}
+          type={'category'}
+          isUpdate
+          value={categoryName}
+          setName={setCategoryName}
+          submit={handleSubmit}
+          messages={messages}
+        />
+      ) : (
+        <Modal
+          trigger={modalType === 'delete'}
+          setTrigger={setModalType}
+          type={'category'}
+          isDelete
+          submit={handleSubmit}
+          messages={messages}
+          name={categoryName}
+        />
+      )}
+
+      {messages.join('').includes('successful') && modalType === '' && (
+        <Modal
+          title={messages}
+          setTrigger={setModalType}
+          trigger={messages.join('').includes('successful') && modalType === ''}
+          notification
+        />
+      )}
     </>
   )
 }
